Do not emit job spans for single queue add
On the context set by instrumented user-facing functions to add jobs
(`(Queue|FlowProducer).add(Bulk)?`), which is propagated to the also
instrumented `Job.add`, set a flag so that `Job.add` can determine
if it's being called from a bulk context (an `.addBulk` function,
which will invoke it several times with different jobs to be added
in parallel) and/or from a flow context (a `FlowProducer.` method,
which will invoke it several time with a parent job and its children)

When it's being called from either context, `Job.add` will emit a
span for each job added by the bulk/flow context. When it is not
being called, `Job.add` will not emit any spans, with the span
emitted by `Queue.add` being the producer kind span.