Allow config to require producer parent spans
Implement a `requireParentSpanForProducers` config option, which
defaults to `false`. When set to `true`, producer spans (those of
the form `(Queue|FlowProducer).add(Bulk)?`, and the `Job.add` spans
that they spawn as child spans) will not be emitted unless the
current active context has a span -- that is, unless the emitted
span would have a parent.

This avoids the situation where the act of enqueueing a job in an
uninstrumented context becomes a trace in itself. Consumer spans
(representing the processing of a previously enqueued job) are
not affected by this config option, and are always emitted regardless
of whether they have a parent span.