Align with Messaging Semantic Conventions
- Use `process`, not `receive`, for the `messaging.operation` value
in the `Worker.run` span.
- Remove `messaging.bullmq.(worker|queue).name` attributes. The name
of the worker is the name of the queue, which is represented in the
`messaging.destination` attributes.
- Use a `messaging.bullmq.operation.name` to provide BullMQ-meaningful
names of operations. This mirrors the `messaging.operation.name`
attribute in the not-yet-published v1.26.0 of the OpenTelemetry
Semantic Conventions.
- Name spans by the destination (the queue) and the operation type.
Notably, this removes the job name, which per BullMQ's documentation,
is meant to be human-meaningful and uniquely identify the job, meaning
it's very likely to be high-cardinality.