Tweak "producer as parent" config option
Some tweaks to the changes introduced in #11.

Rename the config option to `useProducerSpanAsConsumerParent` (that
is, `s/Context/Span`) to use language that is more familiar to casual
OpenTelemetry users.

Rewrite the config option description to describe what happens with
and without the config option.

Rearrange the instrumentation to move logic branches out of the span
creation function.